Simplified some things:
GetChildMutable is gone, replaced with an advanced GetChild template where the child assumes the same constness as the parent. This typically works for the access patterns we want:

if we're searching, we acquire read locks along the path down the root. If we're inserting or removing, we acquire write locks along the path from the root.