SplitNode => SplitLeafNode and SplitInternalNode.
Both take in mutable nodes instead of const nodes and no longer clone. They assume they can modify the left node in place.

The callers now Clone before splitting.

The recursive call of InsertIntoLeaf is guaranteed to not clone and not split. InsertIntoLeaf now has a const and non-const implementation. It only clones if it takes in a const leaf.