Delete version stuff from BlobStore because it makes more sense to stick it to BPlusTree.
Introduce a BPlusTreeHeader that has a version and an index to the root node.

TODO: Introduce a CompareAndSwap of some sort to BlobStoreObject, cloning the header every time and incrementing the version and pointing to the previous header.

We probably also want the header version to be used on every subsequent post-clone operation.