add_chunk => get_or_create_chunk.
This is VERY subtle.

Multiple processes using SharedMemoryAllocator concurrently could realize they need more memory and stampede toward adding chunks which might end up allocating FAR more memory than necessary.

Instead when we realize we need more memory we ask for the next chunk index after the one currently known.

Only one thread/process will get to allocate a new chunk or multiple chunks and that process can be the one that adds a new free node for the chunk.

Note this isn't strictly true due to the ABA problem, the num_chunks quantity is not monotonically increasing. It can go up and down. That means that we can remove a chunk and readd it again in between test and update, and then we might succeed to update, and report that.