This should fix some race conditions in SharedMemoryAllocator:
AllocatedNodeHeader and FreeNodeHeader are now a single node type: Node.

The transition between Allocated and Free won't touch things like chunk_index so we'll always safely grab the chunk index. We do however also track the node's version number which is updated on state transitions between allocated and unallocated.

We currently have a separate NodeType enum, but it's conceivable and perhaps even desirable that the version counter also indicates free or allocated with allocated being an odd number and free being even.