A transaction cannot assume a HeadNode at a certain location. Instead that's a parameter passed into the transaction.
Who creates the head node? Does it need a root index?

One possibility is all objects are named. The name needs to be stable.

Perhaps we have a B+ tree with name as the keys and the index of the head node of the transactional data structure as the value.