Track mutated objects in a separate map. This is another step towards being able to merge transactions. Realistically, one or more transactions will commit before another is able to.
We need to squash all transactions above the old_head into one virtual transaction, and then merge the given transaction on top of it.

We can compare mutations and see what it would take to apply the transaction on top.

For files, the structure and the data are tightly interwined. For trees, we don't care about the structure, just the key value pairs.